+++ api/controllers/teams.go
@@ -1,6 +1,7 @@
 package controllers
 
 import (
+	"api/assertions"
 	"api/services/teams_service"
 	"api/services/users_service"
 	"api/structs"
@@ -39,9 +40,15 @@ func GetTeamById(c *gin.Context) {
 		return
 	}
 
-	// TODO: ADD BELOW LOGIC
-	// If current user is not member of team and is not admin,
-	// throw an error
+	data, _ := c.Get("authScope")
+	authScope := data.(structs.AuthScope)
+
+	membership_err := assertions.UserIsMemberOfTeam(team.ID, uint(authScope.UserID))
+
+	if membership_err != nil {
+		c.AbortWithStatusJSON(http.StatusForbidden, gin.H{"status": "error", "message": "User not a member of the specified team.", "data": nil})
+		return
+	}
 
 	// Return team info
 	c.JSON(http.StatusOK, gin.H{"status": "success", "message": "Team found.", "data": team})
@@ -100,29 +107,17 @@ func DeleteTeam(c *gin.Context) {
 		return
 	}
 
-	// TODO: ADD BELOW LOGIC
-	// If user is not within the team, is not of type manager on
-	// the current team or is not an admin, throw an error
+	data, _ := c.Get("authScope")
+	authScope := data.(structs.AuthScope)
+
+	team_manager_err := assertions.UserIsManagerOfTeam(team_to_delete.ID, uint(authScope.UserID))
+
+	if team_manager_err != nil {
+		c.AbortWithStatusJSON(http.StatusForbidden, gin.H{"status": "error", "message": "You do not have permission to perform that action.", "data": nil})
+		return
+	}
 
 	deleted_team, _ := teams_service.DeleteTeam(team_id)
 
 	c.JSON(http.StatusOK, gin.H{"status": "success", "message": "Team successfully deleted.", "data": deleted_team})
 }
-
-// func GetAllTeamApplications(c *gin.Context) {
-// 	team_input_param := c.Param("team_id")
-// 	team_id, conv_err := strconv.Atoi(team_input_param)
-
-// 	if conv_err != nil {
-// 		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"status": "error", "message": "Error requesting team by id.", "data": nil})
-// 		return
-// 	}
-
-// 	// TODO: ADD BELOW LOGIC
-// 	// If user is not within the team, is not of type manager on
-// 	// the current team or is not an admin, throw an error
-
-// 	applications := applications_service.GetAllApplications(models.Application{TeamID: uint(team_id)})
-
-// 	c.JSON(http.StatusOK, gin.H{"status": "success", "message": "Successfully retrieved applications for team.", "data": gin.H{"applications": applications}})
-// }
